VERSAO FINAL 1.0

+++ TrabalhoFinal/TrabalhoFinalLPOO2/src/model/dao/interfaces/LocacaoDAO.java
@@ -7,7 +7,7 @@ import misc.Locacao;
 
 public interface LocacaoDAO {
     public List<Locacao> getAll()throws SQLException, IOException;
-    public Locacao getById(int id) throws SQLException, IOException;
+    public Locacao getById(long id) throws SQLException, IOException;
     public void add(Locacao locacao) throws SQLException, IOException;
     public void update(Locacao locacao) throws SQLException, IOException;
     public void delete(Locacao locacao) throws SQLException, IOException;
